pyproject.toml
[tool.pytest.ini_options]
testpaths = "tests"
norecursedirs = ".git"
asyncio_mode = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
show_error_codes = true
follow_imports = "silent"
local_partial_types = true
strict_equality = true
no_implicit_optional = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
warn_return_any = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "homeassistant.components.*"
ignore_missing_imports = true

[tool.coverage.paths]
source = ["custom_components/audiobookshelf", "utils", "tests"]

[tool.ruff]
select = ["E", "W", "F", "B", "C", "I", "N", "UP", "ANN", "BLE", "A", "COM", "C4", "G", "ISC", "NPY", "INP", "PIE", "T20", "PYI", "Q", "RET", "SLF", "SIM", "TID", "RUF"]
ignore = ["ANN101"]
line-length = 150
target-version = "py310"
exclude = ["utils/measure/playground", "utils/visualize", ".github", "docs"]

[tool.ruff.per-file-ignores]
"utils/measure/**.py" = ["T201"]

[tool.ruff.mccabe]
max-complexity = 15

[tool.poetry]
name = "audiobookshelf"
version = "1"
description = "Audiobookshelf HA custom component"
authors = ["wolffshots <16850875+wolffshots@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3"
ruff = "^0.0"
voluptuous-stubs = "^0.1"
homeassistant-stubs = "2023.8.1"
numpy = "^1.23"
black = "^23.3"
freezegun = "^1.2"
pytest-homeassistant-custom-component = "^0.13"
types-pytz = "^2023.3"
types-croniter = "^1.3"
croniter = "^1.4"
pytest-cov = "^4.0"
coverage = "^7.2"

[tool.poetry.dependencies]
python = "^3.11"
asynctest = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"